에러 잡기!

+++ Assets/Script/EnemyManagerScript/EnemyBase.cs
@@ -5,7 +5,7 @@ namespace Script.EnemyManagerScript
     public class EnemyBase : MonoBehaviour
     {
         protected string EnemyName;
-        protected int HealthPoint;
+        protected float HealthPoint;
         protected internal int CrushDamage;
         protected internal float MoveSpeed;
         public enum EnemyZone { A,B,C,D,E }
